chore: move dom-styles-reader as it is not a mixing anymore
