+++ lib/minimap-element.js
@@ -5,7 +5,7 @@ import { EventsDelegation, AncestorsMethods } from 'atom-utils-plus'
 import elementResizeDetectorImport from 'element-resize-detector'
 
 import * as Main from './main'
-import DOMStylesReader from './mixins/dom-styles-reader'
+import DOMStylesReader from './dom-styles-reader'
 import CanvasDrawer from './mixins/canvas-drawer'
 import include from './decorators/include'
 import element from './decorators/element'
